requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
asgiref==3.5.2
asyncmy==0.2.5
black==22.6.0
click==8.1.3
colorama==0.4.4
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
iso8601==1.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pydantic==1.9.1
PyJWT==2.4.0
PyMySQL==1.0.2
pypika-tortoise==0.1.5
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
tomli==2.0.1
tortoise-orm==0.19.1
typing-extensions==4.2.0
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
